fix(kieu): dong 5 nhom loi mypy co the la bug that, 79 -> 0 (LOW-308)
`mypy --ignore-missing-imports --follow-imports=silent`: 262 loi / 71 tep. Nam
nhom ticket chon — arg-type 15, union-attr 11, return-value 9, attr-defined 24,
index 20 — nay ve 0. Tong con 172. Phan con lai (var-annotated 91, assignment 59)
la thieu chu thich kieu, khong doi hanh vi, CO Y de lai. Chua dua mypy vao CI.

HAI BUG THAT (test do-truoc-xanh-sau, tests/test_low308_type_bugs.py):

1. article_extract: `soup.title.string` la None khi `<title>` RONG, `.strip()`
   tren None giet ca lan boc. Day la duong LUI khi trang khong co og:title —
   dung luc trang da kho boc — va nguoi goi (`material.py`) chi thay tien trinh
   con thoat khac 0. Doi sang `get_text(strip=True)`, khong bao gio tra None.
2. gin_submit: `cv2.imread` tra None khi tep anh khong con/khong doc duoc; None
   di tiep vao `use_mask` roi no ra `AttributeError: 'NoneType' has no attribute
   'shape'` o swap_image_text.py:107. Nay noi thang tep nao va chay lai lenh gi.

CHU THICH KIEU SAI (docstring da noi dung, chu ky thi khong — 0 dong chay doi):
`_handle_lock -> bool` (docstring: "Tra ve True khi gap KHOA MO COI"; cai sai nay
keo theo loi thu hai o `count_crashes`), `_drop_gate_old_chat -> str | None`,
`_ask_api -> dict | None` (docstring: "None ... KHONG phai {} rong"),
`check_guide_source_compact`/`check_quote_translated`/`check_empty_image` nhan
`| None` (ca ba DA tu xu ly None: `(chu or "")`, `if not a`),
`count_image_use_ok(anh: list | None)`, `height_item -> float`.

TEN BI DUNG LAI TRONG CUNG MOT HAM (cai nay giau bug that su):
- journal.py: `gy` vua la danh sach giay vua la chuoi, `d` vua la danh sach draft
  vua la mot dong chu -> `giay` / `dong`.
- scan_models.py: `r` vua la Response httpx vua la mot hang CSV -> `hang`.
- ranking.py: `i` vua la chi so vua la mot anh PIL -> `anh`.

PILLOW: `Image.LANCZOS`/`BOX` la bi danh DA CU (Pillow >= 9.1 dung
`Image.Resampling.*`); stub khong con khai nen mypy bao attr-defined. Doi 27 cho.
DA DO, khong doan: dung 3 the that qua card.build (ba ti le 16:9, 4:5, 1:1) o ban
cu va ban moi -> md5 GIONG HET ca ba; dhash cua ca ba anh cung giong het. Khong
mot pixel nao doi.

CON LAI: `cv2.ORB_create` CO THAT luc chay, chi thieu trong stub -> mot
`# type: ignore[attr-defined]` kem ly do (khong phai ignore tran). `reconfigure`
cua stdout hoi bang `getattr` thay vi de `except Exception` nuot — stdout co the
da bi thay bang StringIO trong test. Cac `dict[str,X].get(<co the None>)` chay
dung nhung khai ro bang `or ""`. Cac dict/list long nhieu kieu khai `: dict` /
`: list[dict]` de may doc kieu khong chi thay kieu cua khoa dau tien.

Kiem: 5 nhom = 0; tests/run.sh 155/155; pyflakes sach.

+++ about_text.py
@@ -37,7 +37,7 @@ def font_default(h_vung: int, h_anh: int) -> str:
     return "bold" if h_vung >= 0.045 * h_anh else "regular"
 
 
-def height_item(d: ImageDraw.ImageDraw, text: str, font) -> int:
+def height_item(d: ImageDraw.ImageDraw, text: str, font) -> float:
     """Chieu cao MUC that cua mot dong khi ve (bbox chu, khong phai bbox font)."""
     bb = d.textbbox((0, 0), text, font=font)
     return max(1, bb[3] - bb[1])
